# Conflicts:
#	Desarrollo/BulletTest/.vs/BulletTest/v14/.suo
#	Desarrollo/BulletTest/BulletTest.VC.db
#	Desarrollo/BulletTest/BulletTest/Debug/AnimatedSceneNode.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Animation.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Asalto.obj
#	Desarrollo/BulletTest/BulletTest/Debug/AsaltoDrop.obj
#	Desarrollo/BulletTest/BulletTest/Debug/BasicSceneNode.obj
#	Desarrollo/BulletTest/BulletTest/Debug/BulletTest.log
#	Desarrollo/BulletTest/BulletTest/Debug/BulletTest.obj
#	Desarrollo/BulletTest/BulletTest/Debug/BulletTest.tlog/CL.read.1.tlog
#	Desarrollo/BulletTest/BulletTest/Debug/BulletTest.tlog/CL.write.1.tlog

#	Desarrollo/BulletTest/BulletTest/Debug/BulletTest.tlog/link.read.1.tlog
#	Desarrollo/BulletTest/BulletTest/Debug/Camera.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Cliente.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Clock.obj
#	Desarrollo/BulletTest/BulletTest/Debug/ClockImpl.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Enemy.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Entity.obj
#	Desarrollo/BulletTest/BulletTest/Debug/EntityManager.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Game.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Granada.obj
#	Desarrollo/BulletTest/BulletTest/Debug/GraphicEngine.obj
#	Desarrollo/BulletTest/BulletTest/Debug/GunBullet.obj
#	Desarrollo/BulletTest/BulletTest/Debug/InputHandler.obj

#	Desarrollo/BulletTest/BulletTest/Debug/KinematicCharacterController.obj
#	Desarrollo/BulletTest/BulletTest/Debug/LifeObject.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Lista.obj
#	Desarrollo/BulletTest/BulletTest/Debug/MastEventReceiver.obj
#	Desarrollo/BulletTest/BulletTest/Debug/MessageHandler.obj
#	Desarrollo/BulletTest/BulletTest/Debug/PhysicsEngine.obj
#	Desarrollo/BulletTest/BulletTest/Debug/PhysicsEntity.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Pistola.obj
#	Desarrollo/BulletTest/BulletTest/Debug/PistolaDrop.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Player.obj
#	Desarrollo/BulletTest/BulletTest/Debug/RenderState.obj
#	Desarrollo/BulletTest/BulletTest/Debug/RocketBullet.obj
#	Desarrollo/BulletTest/BulletTest/Debug/RocketLauncher.obj
#	Desarrollo/BulletTest/BulletTest/Debug/RocketLauncherDrop.obj
#	Desarrollo/BulletTest/BulletTest/Debug/SceneNode.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Source.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Time.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Util.obj
#	Desarrollo/BulletTest/BulletTest/Debug/Weapon.obj
#	Desarrollo/BulletTest/BulletTest/Debug/WeaponDrop.obj
#	Desarrollo/BulletTest/BulletTest/Debug/main.obj
#	Desarrollo/BulletTest/BulletTest/Debug/stdafx.obj
#	Desarrollo/BulletTest/BulletTest/Debug/vc140.idb
#	Desarrollo/BulletTest/BulletTest/Debug/vc140.pdb
#	Desarrollo/BulletTest/BulletTest/Entities/Granada.cpp
#	Desarrollo/BulletTest/BulletTest/Entities/Weapons/Asalto.cpp
#	Desarrollo/BulletTest/Debug/BulletTest.exe
#	Desarrollo/BulletTest/Debug/BulletTest.ilk
#	Desarrollo/BulletTest/Debug/BulletTest.pdb
#	Desarrollo/BulletTest/Debug/RaknetServerv0.1.exe
#	Desarrollo/BulletTest/Debug/RaknetServerv0.1.ilk
#	Desarrollo/BulletTest/Debug/RaknetServerv0.1.pdb
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/Clock.obj
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/ClockImpl.obj
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/Entity.obj
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/EntityManager.obj
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/MainServer.obj
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/Player.obj
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/RaknetServerv0.1.log

#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/RaknetServerv0.1.tlog/CL.command.1.tlog

#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/RaknetServerv0.1.tlog/CL.read.1.tlog

#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/RaknetServerv0.1.tlog/CL.write.1.tlog

#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/RaknetServerv0.1.tlog/RaknetServerv0.1.lastbuildstate

#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/RaknetServerv0.1.tlog/link.command.1.tlog

#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/RaknetServerv0.1.tlog/link.read.1.tlog

#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/RaknetServerv0.1.tlog/link.write.1.tlog
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/RenderState.obj
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/Time.obj
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/vc140.idb
#	Desarrollo/BulletTest/RaknetServerv0.1/Debug/vc140.pdb

+++ Desarrollo/BulletTest/BulletTest/Game.cpp
@@ -124,7 +124,7 @@ void Game::inicializar()
 
 	PhysicsEntity *sueloEnt2 = new PhysicsEntity(suelo2, "suelo2");
 	sueloEnt2->setRigidBody(PhysicsEngine::i().createBoxRigidBody(sueloEnt2, Vec3<float>(100, 5.f, 100), 0));
-	sueloEnt2->rotate(Vec3<float>(90 * PI / 180.0,0, 0));
+	sueloEnt2->rotate(Vec3<float>(float(90 * PI / 180.0),0, 0));
 	sueloEnt2->setPosition(Vec3<float>(-50, 0, 0));
 	///////////////////////////////////////////////////////////////
 
@@ -132,7 +132,7 @@ void Game::inicializar()
 
 	PhysicsEntity *sueloEnt3 = new PhysicsEntity(suelo3, "suelo3");
 	sueloEnt3->setRigidBody(PhysicsEngine::i().createBoxRigidBody(sueloEnt3, Vec3<float>(100, 5.f, 100), 0));
-	sueloEnt3->rotate(Vec3<float>(0, 0, 90 * PI / 180.0));
+	sueloEnt3->rotate(Vec3<float>(0, 0, float( 90 * PI / 180.0)));
 	sueloEnt3->setPosition(Vec3<float>(0, 0, -50));
 
 
@@ -142,7 +142,7 @@ void Game::inicializar()
 
 	PhysicsEntity *sueloEnt4 = new PhysicsEntity(suelo4, "suelo4");
 	sueloEnt4->setRigidBody(PhysicsEngine::i().createBoxRigidBody(sueloEnt4, Vec3<float>(25, 5.f, 100), 0));
-	sueloEnt4->rotate(Vec3<float>(0, 0, 30 * PI / 180.0));
+	sueloEnt4->rotate(Vec3<float>(0, 0, float(30 * PI / 180.0)));
 	sueloEnt4->setPosition(Vec3<float>(60, -25, 40));
 
 	///////////////////////////////////////////////////////////////////////
@@ -161,7 +161,7 @@ void Game::inicializar()
 
 	PhysicsEntity *sueloEnt6 = new PhysicsEntity(suelo6, "suelo6");
 	sueloEnt6->setRigidBody(PhysicsEngine::i().createBoxRigidBody(sueloEnt6, Vec3<float>(100, 5.f, 100), 0));
-	sueloEnt6->rotate(Vec3<float>(90 * PI / 180.0, 0, 0));
+	sueloEnt6->rotate(Vec3<float>(float(90 * PI / 180.0), 0, 0));
 	sueloEnt6->setPosition(Vec3<float>(75, 0, 0));
 	///////////////////////////////////////////////////////////////
 
@@ -169,7 +169,7 @@ void Game::inicializar()
 
 	PhysicsEntity *sueloEnt7 = new PhysicsEntity(suelo7, "suelo7");
 	sueloEnt7->setRigidBody(PhysicsEngine::i().createBoxRigidBody(sueloEnt7, Vec3<float>(100, 5.f, 100), 0));
-	sueloEnt7->rotate(Vec3<float>(0, 0, 90 * PI / 180.0));
+	sueloEnt7->rotate(Vec3<float>(0, 0, float( 90 * PI / 180.0)));
 	sueloEnt7->setPosition(Vec3<float>(0, 0, 50));
 
 	//////////////////////////////////////////////////////////////////////
